# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateEvidenceEnvelope": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/EvidenceCandidate"
          },
          "counterexamples": {
            "$ref": "#/components/schemas/EvidenceCounterexamples"
          },
          "evidence_level": {
            "const": "candidate",
            "title": "Evidence Level"
          },
          "independence": {
            "$ref": "#/components/schemas/EvidenceIndependence"
          },
          "ledger": {
            "$ref": "#/components/schemas/EvidenceLedger"
          },
          "result": {
            "$ref": "#/components/schemas/EvidenceResult"
          },
          "schema_version": {
            "const": "evidence-envelope-v1",
            "title": "Schema Version"
          },
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          }
        },
        "required": [
          "schema_version",
          "evidence_level",
          "candidate",
          "source",
          "result",
          "independence",
          "counterexamples",
          "ledger"
        ],
        "title": "CandidateEvidenceEnvelope",
        "type": "object"
      },
...
      "CandidateReference": {
        "additionalProperties": false,
        "properties": {
          "candidate_note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Note"
          },
          "description": {
            "maxLength": 600,
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "maxLength": 120,
            "title": "Domain",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/CandidateEvidenceEnvelope"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retrieval_rank": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Retrieval Rank",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "domain",
          "description",
          "retrieval_rank",
          "evidence"
        ],
        "title": "CandidateReference",
        "type": "object"
      },
...
      "DiagnoseRequest": {
        "additionalProperties": false,
        "properties": {
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 12,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{11,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "situation": {
            "maxLength": 1500,
            "minLength": 1,
            "title": "Situation",
            "type": "string"
          }
        },
        "required": [
          "situation"
        ],
        "title": "DiagnoseRequest",
        "type": "object"
      },
      "DiagnoseResponse": {
        "additionalProperties": false,
        "properties": {
          "assessment_kind": {
            "const": "structural_state_hypothesis",
            "title": "Assessment Kind"
          },
          "candidate_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_version": {
            "const": "secondary-tools-v2",
            "title": "Contract Version"
          },
          "evidence": {
            "$ref": "#/components/schemas/CandidateEvidenceEnvelope"
          },
          "evolution": {
            "maxLength": 1200,
            "minLength": 1,
            "title": "Evolution",
            "type": "string"
          },
          "primary_state": {
            "$ref": "#/components/schemas/StructuralState"
          },
          "reasoning": {
            "maxLength": 1500,
            "minLength": 1,
            "title": "Reasoning",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Recommendations",
            "type": "array"
          },
          "request_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{11,63}$",
            "title": "Request Id",
            "type": "string"
          },
          "secondary_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuralState"
              },
              {
                "type": "null"
              }
            ]
          },
          "signals_to_watch": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "title": "Signals To Watch",
            "type": "array"
          },
          "situation": {
            "maxLength": 1500,
            "minLength": 12,
            "title": "Situation",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "request_id",
          "situation",
          "assessment_kind",
          "primary_state",
          "secondary_state",
          "reasoning",
          "evolution",
          "signals_to_watch",
          "recommendations",
          "candidate_reference",
          "evidence"
        ],
        "title": "DiagnoseResponse",
        "type": "object"
      },
...
      "EvidenceCandidate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "score": {
            "title": "Score",
            "type": "null"
          },
          "status": {
            "const": "recorded",
            "title": "Status"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceCandidate",
        "type": "object"
      },
      "EvidenceCounterexamples": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": [
              "not_recorded",
              "gap_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "status"
        ],
        "title": "EvidenceCounterexamples",
        "type": "object"
      },
      "EvidenceIndependence": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "not_recorded",
            "title": "Kind"
          },
          "status": {
            "const": "not_recorded",
            "title": "Status"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceIndependence",
        "type": "object"
      },
      "EvidenceLedger": {
        "additionalProperties": false,
        "properties": {
          "artifact_sha256": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Sha256"
          },
          "claim_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Id"
          },
          "recorded_at": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "status": {
            "const": "not_recorded",
            "title": "Status"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "version": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "status"
        ],
        "title": "EvidenceLedger",
        "type": "object"
      },
      "EvidenceResult": {
        "additionalProperties": false,
        "properties": {
          "provenance": {
            "enum": [
              "NOT_TESTED",
              "INTERNAL_AI_SCREEN"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "status": {
            "enum": [
              "recorded",
              "not_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "verdict": {
            "enum": [
              "NOT_TESTED",
              "INCONCLUSIVE"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "status",
          "provenance",
          "verdict"
        ],
        "title": "EvidenceResult",
        "type": "object"
      },
      "EvidenceSource": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "not_recorded",
              "internal_kb"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "source_review": {
            "title": "Source Review",
            "type": "null"
          },
          "status": {
            "enum": [
              "recorded",
              "not_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "null"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceSource",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchRequest": {
        "properties": {
          "lang": {
            "default": "zh",
            "description": "Output language for LLM-generated text: 'zh' or 'en'",
            "title": "Lang",
            "type": "string"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "rewrite": {
            "default": false,
            "description": "Use LLM to rewrite query for better matching",
            "title": "Rewrite",
            "type": "boolean"
          },
          "top_k": {
            "default": 12,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
...
      "StructuralState": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Definition",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "state_id": {
            "pattern": "^[a-z][a-z0-9_]{2,63}$",
            "title": "State Id",
            "type": "string"
          },
          "typical_signal": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Typical Signal",
            "type": "string"
          }
        },
        "required": [
          "state_id",
          "name",
          "definition",
          "typical_signal"
        ],
        "title": "StructuralState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Optional API key. Anonymous requests fall back to the free tier (60 req/min). Provide an X-API-Key header to access higher tiers.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Structural Team",
      "url": "https://structural.bytedance.city/"
    },
    "description": "跨领域结构同构搜索引擎 — Structural Isomorphism API.\n\nPublic surface for the Structural cross-domain knowledge engine. Endpoints are grouped by capability (Ask / Phase / Phenomenon / Pricing / Newsletter / Admin). Rate limits are tier-based; supply an `X-API-Key` header to promote a request beyond the anonymous free-tier bucket.",
    "license": {
      "name": "Proprietary"
    },
    "title": "Structural",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/diagnose": {
      "post": {
        "description": "Run a structural diagnosis on one organisation/team situation.",
        "operationId": "diagnose_api_diagnose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnoseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnose",
        "tags": [
          "diagnose"
        ]
      }
    },
...
    "/api/search": {
      "post": {
        "operationId": "search_phenomena_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Phenomena",
        "tags": [
          "search"
        ]
      }
    },
...
  },
...
}